+++ src/lib/config/config.ts
@@ -111,7 +111,12 @@ let conf =  {
 
     },
     mode:"xmpp",
-    "concurrentRequests": 10,
+    "concurrentRequests": 1000,
+    "requestsRate": {
+        "maxReqByIntervalForRequestRate": 600, // nb requests during the interval.
+        "intervalForRequestRate": 60, // nb of seconds used for the calcul of the rate limit.
+        "timeoutRequestForRequestRate": 600, // nb seconds Request stay in queue before being rejected if queue is full.
+    },
     debug:true,
     permitSearchFromPhoneBook:true,
     displayOrder:"firstLast",